km: ta: mbedtls_proxy: use REE time for cert gen
Since optee_os core doesn't have an absolute time, only relative, we're
not able to derive the current date using the System (TEE) time and
hence not able to generate a valid certificate for testing. An invalid
certificate will cause the TA to panic.

Use the REE time for certificate generation. Obviously this isn't fully
secure, but since this code path is used for development and testing
only, we allow it and print a warning to let the user know what to do
instead in actual platforms or products.

Signed-off-by: Victor Chong <victor.chong@linaro.org>
Acked-by: Ruchika Gupta <ruchika.gupta@linaro.org>